refactor: migrate service layer to use split Medicine/MedicineInformation entities
- Medicine.java: Added @OneToOne(mappedBy) to MedicineInformation for extended
  details. Fixed medicine_owner column to use ENUM columnDefinition to satisfy
  ddl-auto=validate. Renamed bucketId -> subcategoryId (Integer) to fix
  PropertyReferenceException in repository.

- MedicineRepository.java: Updated findBySubcategoryId signature to Integer.

- IMedicineService.java + MedicineServiceImpl.java: Updated getMedicinesBySubcategory
  to accept Integer. Added missing MedicineInformation import.

- MedicineController.java: Updated endpoint path variable type to Integer.

- BucketServiceImpl.java: Fixed all bucket creation and unavailable item loops
  to access image (photo1) and packing via medicine.getMedicineInformation()
  with null-safe guards.

- MainServiceImpl.java: Updated top/selling medicine DTO population to use
  MedicineInformation for image, manufacturer, and packing fields.

- OrderServiceImpl.java: Fixed populateMedicalDTO and populateOrders to access
  details (packing, medicineLogo, useOf, strip) via MedicineInformation.

- PreOrderServiceImpl.java: Fixed populateMedicalDTO same as OrderServiceImpl.

- VendorServiceImpl.java: Added MedicineInformation import and corrected DTO
  mapping in getVendorByIdOrMedicine to use valid MedicineDTO setters.

CONSTRAINT: spring.jpa.hibernate.ddl-auto=validate is preserved throughout.
No database schema changes were made.

+++ src/main/java/com/trego/api/MedicineController.java
@@ -81,7 +81,7 @@ public class MedicineController {
 
     @GetMapping("/subcategories/{subcategoryId}/medicines")
     public Page<MedicineWithStockAndVendorDTO> getMedicinesBySubcategory(
-            @PathVariable Long subcategoryId,
+            @PathVariable Integer subcategoryId,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size
     ) {
